feat(RHINENG-13856): Create progress stepper (#139)
As part of registration assistant refresh, this PR implements the ProgressStepper and some individual step features.

It does not include actual activation key data or content for the selected OS; but it does include the data necessary to step through the stepper.

To test, replace `<Register />` in `Routes.js` with `<NewRegister />` and uncomment the lazy loaded `NewRegister` component in the same file. Run the app and you should see the new content.

Refer to the mocks in the Jira card for specific design.

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>
Co-authored-by: Adonis Puente <60629070+adonispuente@users.noreply.github.com>